% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epps_test.R
\name{epps.statistic}
\alias{epps.statistic}
\title{Estimates the Epps statistic for normal distribution test.}
\usage{
epps.statistic(y)
}
\arguments{
\item{y}{a time series. Missing values are not allowed}
}
\value{
a real value with the Epps test's statistic
}
\description{
Estimates the Epps statistic minimizing the quadratic loss of the characteristic
function of the process in terms of the first two moments.
}
\details{
This function is the equivalent of \code{Sub} in
\emph{Nieto-Reyes, A., Cuesta-Albertos, J. & Gamboa, F. (2014)}.
}
\examples{
# Generating an stationary arma process
y = arima.sim(100,model = list(ar = 0.3))
epps.statistic(y)

}
\references{
Epps, T.W. (1987). Testing that a stationary time series is Gaussian. \emph{The
Annals of Statistic}. 15(4), 1683-1698.\url{http://www.jstor.org/stable/2336512}.
\code{doi:10.1214/aos/1176350618}

Nieto-Reyes, A., Cuesta-Albertos, J. & Gamboa, F. (2014). A random-projection
based test of Gaussianity for stationary processes. \emph{Computational
Statistics & Data Analysis, Elsevier}, vol. 75(C), pages 124-141.
\url{http://www.sciencedirect.com/science/article/pii/S0167947314000243}
\code{doi:https://doi.org/10.1016/j.csda.2014.01.013}.
}
\seealso{
\code{\link{lobato.statistic}}
}
\author{
Alicia Nieto-Reyes and Asael Alonzo Matamoros.
}
