% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{ggacf}
\alias{ggacf}
\title{\code{acf} plot}
\usage{
ggacf(y, title = NULL)
}
\arguments{
\item{y}{a numerical vector or a time series.}

\item{title}{a string with the plot's title.}
}
\value{
None.
}
\description{
Plot for the auto-correlation function for a univariate time series
}
\examples{
x = rnorm(100)
ggacf(x)

}
\author{
Asael Alonzo Matamoros
}
