% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{normal.test}
\alias{normal.test}
\title{The goodness of fit tests for normal distribution.}
\usage{
normal.test(y,normality="epps",alpha=0.05)
}
\arguments{
\item{y}{A univariate time series}

\item{normality}{A character string naming the desired test for checking gaussian distribution.
Valid values are \code{"epps"} for the Epps, \code{"lobato"} for Lobato and Velasco's,\code{"vavra"} for
the Psaradakis and  Vavra, \code{"rp"} for the random projections, \code{"jb"} for the Jarque and Beras,
\code{"ad"} for Anderson Darling test, and \code{"shapiro"} for the Shapiro-Wilk's test. The default value
is \code{"epps"} test.}

\item{alpha}{Level of the test, possible values range from 0.01 to 0.1. By default \code{alpha = 0.05}
is used}
}
\value{
An h.test class with the main results of normal hypothesis test.
}
\description{
Perform a goodness of fit test to check gaussian distribution in stationarity stochastic process.
}
\details{
Several different tests are available:
\code{"lobato"}, \code{"epps"}, \code{"vavras"} and \code{"rp"} test are for testing normal
distributions in stationary time series. \code{"jb"}, \code{"ad"}, and  \code{"shapiro"} tests
are for random numeric vectors. In all cases, the alternative hypothesis is that \code{y} has
a gaussian distribution. By default, \code{alpha = 0.05} is used to select the more likely
hypothesis.
}
\examples{
#  stationary  ar process
y = arima.sim(100,model = list(ar = 0.3))
normal.test(y) # epps test

# normal random sample
y = rnorm(100)
normal.test(y,normality = "shapiro")

# exponential random sample
y = rexp(100)
normal.test(y,normality = "ad")

}
\references{
Epps, T.W. (1987). Testing that a stationary time series is Gaussian. \emph{The
Annals of Statistic}. 15(4), 1683-1698.\url{http://www.jstor.org/stable/2336512}.
\code{doi:10.1214/aos/1176350618}

Lobato, I., & Velasco, C. (2004). A simple test of normality in time series.
\emph{Journal of econometric theory}. 20(4), 671-689.
\url{doi:10.1017/S0266466604204030}

Psaradakis, Z. & Vavra, M. (2017). A distance test of normality for a wide class
of stationary process. \emph{Journal of Econometrics and Statistics}. 2, 50-60.
\url{http://www.sciencedirect.com/science/article/pii/S2452306216300296}.
\code{doi: https://doi.org/10.1016/j.ecosta.2016.11.005}.

Nieto-Reyes, A., Cuesta-Albertos, J. & Gamboa, F. (2014). A random-projection
based test of Gaussianity for stationary processes. \emph{Computational
Statistics & Data Analysis, Elsevier}, vol. 75(C), pages 124-141.
\url{http://www.sciencedirect.com/science/article/pii/S0167947314000243}
\code{doi:https://doi.org/10.1016/j.csda.2014.01.013}.

Patrick, R. (1982). An extension of Shapiro and Wilk's W test for
normality to large samples. \emph{Journal of Applied Statistics}.
31, 115-124. \code{doi: 10.2307/2347973}.

Cromwell, J. B., Labys, W. C. & Terraza, M. (1994). Univariate Tests for
Time Series Models. \emph{Sage, Thousand Oaks, CA}. 20-22.
}
\seealso{
\code{\link{uroot.test}},\code{\link{seasonal.test}}
}
\author{
Asael Alonzo Matamoros
}
\keyword{distribution}
\keyword{normal}
\keyword{series,}
\keyword{stationarity,}
\keyword{time}
