% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_test.R
\name{rp.test}
\alias{rp.test}
\title{The k random projections Test for normal distribution}
\usage{
rp.test(y,k = 2,FDR = FALSE,pars1 = c(100,1),pars2  = c(2,7),seed = NULL)
}
\arguments{
\item{y}{a numeric vector or an object of the ts class containing an univariate
stationary time series.}

\item{k}{an integer with the number of random projections to be used, by default
k = 2.}

\item{FDR}{a logical value for mixing the p-values using a dependent False discovery
rate method. By default FDR = FALSE.}

\item{pars1}{an optional real vector with the shape parameters of the beta distribution
used for the odd number random projection. By default, pars1 = c(100,1) where,
shape1 = 100 and shape2 = 1.}

\item{pars2}{an optional real vector with the shape parameters of the beta distribution
used for the even number random projection. By default, pars1 = c(2,7) where,
shape1 = 2 and shape2 = 7.}

\item{seed}{An optional \code{\link[=set.seed]{seed}} to use.}
}
\value{
a h.test class with the main results of the Epps hypothesis test. The
h.test class have the following values
\itemize{
 \item{"k"}{The number of used projections}
 \item{"lobato"}{The average Lobato and Velasco's test statistics of the k projected samples}
 \item{"epps"}{The average Epps test statistics of the k projected samples}
 \item{"p.value"}{The mixed p value}
 \item{"alternative"}{The alternative hypothesis}
 \item{"method"}{The used method: rp.test}
 \item{"data.name"}{The data name}
}
}
\description{
Performs the random projections test for normal distribution of a univariate
stationary process. The initial hypothesis (H0), is that the process follows
a normal distribution, and k is the number of used random projections.
}
\details{
The random projections test generates k independent random projections of the process. A Lobato and
Velasco's test is applied to the first half of the projections. And a epps test for
the other half. By default, a Monte Carlo p.value estimate is used for mixing the
tests. A False discovery rate can be used for mixing by setting FDR = TRUE.

For generating the k random projections a beta distribution is used. By default a
beta(shape1 = 100,shape = 1) and a beta(shape1 = 2,shape = 7) are used to generate
the odd and even projections respectively. For using a different parameter set, change
pars1 or pars2.

The test was proposed by \emph{Nieto-Reyes, A.,Cuesta-Albertos, J. & Gamboa, F. (2014)}.
}
\examples{
# Generating an stationary arma process
y = arima.sim(100,model = list(ar = 0.3))
rp.test(y)

}
\references{
Nieto-Reyes, A., Cuesta-Albertos, J. & Gamboa, F. (2014). A random-projection
based test of Gaussianity for stationary processes. \emph{Computational
Statistics & Data Analysis, Elsevier}, vol. 75(C), pages 124-141.
\url{http://www.sciencedirect.com/science/article/pii/S0167947314000243}
\code{doi:https://doi.org/10.1016/j.csda.2014.01.013}.

Epps, T.W. (1987). Testing that a stationary time series is Gaussian. \emph{The
Annals of Statistic}. 15(4), 1683-1698.\url{http://www.jstor.org/stable/2336512}.
\code{doi:10.1214/aos/1176350618}

Lobato, I., & Velasco, C. (2004). A simple test of normality in time series.
\emph{Journal of econometric theory}. 20(4), 671-689.
\url{doi:10.1017/S0266466604204030}
}
\seealso{
\code{\link{lobato.test}} \code{\link{epps.test}}
}
\author{
Asael Alonzo Matamoros
}
