% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{check_plot.ts}
\alias{check_plot.ts}
\alias{check_plot}
\alias{check_plot.arima0}
\alias{check_plot.Arima}
\alias{check_plot.fGarch}
\alias{check_plot.numeric}
\alias{check_plot.lm}
\alias{check_plot.HoltWinters}
\alias{check_plot.ets}
\alias{check_plot.forecast}
\title{Generic function for a visual check of residuals in time series models.}
\usage{
\method{check_plot}{ts}(y, model = " ", ...)
}
\arguments{
\item{y}{a numeric vector or an object of the \code{ts} class containing a stationary time series.}

\item{model}{A string with the model name.}

\item{...}{Other plotting parameters to affect the plot.}
}
\value{
A graph object from ggplot2.
}
\description{
Generic function for a visual check of residuals in time series models, these methods are inspired in
the \code{check.residuals} function provided by the \code{forecast} package.
}
\examples{
y = arima.sim(100,model = list(ar = 0.3))
check_plot(y)


}
\seealso{
\code{check_residuals}
}
\author{
Asael Alonzo Matamoros.
}
